team_service/requirements.txt
Django==4.2
djangorestframework==3.14
gunicorn==21.2.0
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1